requirements.txt
Django==3.1.8
asgiref==3.3.4
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
drf-nested-routers==0.93.3
